pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arxiv-rag"
version = "1.0.0"
description = "arXiv RAG System - Search and analyze scientific documents"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "arXiv RAG Team"},
]
keywords = ["arxiv", "rag", "llm", "scientific-papers", "vector-search"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "arxiv>=2.1.0",
    "click>=8.1.7",
    "langchain>=0.1.0",
    "chromadb>=0.4.18",
    "openai>=1.0.0",
    "pypdf2>=3.0.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.5.1",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
arxiv-rag = "cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
